Updated to VS2022, learned vspkg to get CppAD (along with manually adding ipopt) to get non-linear curve fitting to work.  Improved tab loading to now be on demand with a "+" tab to speed loading and prevent unused SQL connections.  Enabled loading blob data from SQL with base64pp library.  Added support for Lifetime graphing.

+++ CV_Theoretical.cpp
@@ -89,7 +89,7 @@ private:
 				double dr = 1;
 
 				// Find zero
-				Evaluation eval( cos( M_PI * (i - 0.25) / (eDEGREE + 0.5) ) );
+				Evaluation eval( cos( M_PI * (i - 0.25) / (double(eDEGREE) + 0.5) ) );
 				do
 				{
 					dr = eval.v() / eval.d();
